Create DI for remote api and commented out the code in RetrofitBuilder
