+++ app/src/main/java/com/kenchen/capstonenewsapp/networking/RetrofitBuilders.kt
@@ -1,27 +1,24 @@
 package com.kenchen.capstonenewsapp.networking
 
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-
-fun buildClient(): OkHttpClient =
-    OkHttpClient.Builder()
-        .addInterceptor(HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        })
-        .build()
-
-fun buildRetrofit(): Retrofit {
-    return Retrofit.Builder()
-        .client(buildClient())
-        .baseUrl(BASE_URL)
-        .addConverterFactory(MoshiConverterFactory.create().asLenient())
-        .build()
-}
-
-fun buildApiService(): RemoteApiService {
-    return buildRetrofit().create(RemoteApiService::class.java)
-}
-
-
+//const val BASE_URL = "https://newsapi.org/v2/"
+//
+//fun buildClient(): OkHttpClient =
+//    OkHttpClient.Builder()
+//        .addInterceptor(HttpLoggingInterceptor().apply {
+//            level = HttpLoggingInterceptor.Level.BODY
+//        })
+//        .build()
+//
+//fun buildRetrofit(): Retrofit {
+//    return Retrofit.Builder()
+//        .client(buildClient())
+//        .baseUrl(BASE_URL)
+//        .addConverterFactory(MoshiConverterFactory.create().asLenient())
+//        .build()
+//}
+//
+//fun buildApiService(): RemoteApiService {
+//    return buildRetrofit().create(RemoteApiService::class.java)
+//}
+//
+//
